Create separate files for baserepo and bikerepo
