+++ api/dependencies/repository_factory.py
@@ -7,15 +7,15 @@ from typing import TypeVar
 from fastapi import Depends
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from api.db import database, repository
+from api.db import database, repository_base
 from api.models import db_models
 
 Model = TypeVar("Model", bound=db_models.Base)
-Repo = TypeVar("Repo", bound=repository.DatabaseRepository)
+Repo = TypeVar("Repo", bound=repository_base.DatabaseRepository)
 
 
 def get_repository(
-    model: type[Model], repository_class: type[Repo] = repository.DatabaseRepository
+    model: type[Model], repository_class: type[Repo] = repository_base.DatabaseRepository
 ) -> Callable[[AsyncSession], Repo]:
     """Get a repository instance for a model.
 
@@ -28,7 +28,7 @@ def get_repository(
     """
 
     def func(session: AsyncSession = Depends(database.get_db_session)) -> Repo:
-        if repository_class == repository.DatabaseRepository:
+        if repository_class == repository_base.DatabaseRepository:
             return repository_class(model, session)
         return repository_class(session)
